Numerous grammatical improvements
I did not realize this is so big. I hope I understand the material well enough to improve the understanding and to make it sound better. I edited less than half; up to "Binding and BindingExpression". This editor is funny; it jumps around sometimes. I hope I did not edit something that I did not know I changed.